[support] Add basic support of collecting info on Windows
