+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Changelog
 
-### [13.0.1](https://kaos.sh/ek/13.0.1)
+### [13.1.0](https://kaos.sh/ek/13.1.0)
 
 - `[env]` Fixed compatibility with Windows
+- `[support]` Add basic support of collecting info on Windows
 
 ### [13.0.0](https://kaos.sh/ek/13.0.0)
