+++ ek.go
@@ -21,7 +21,7 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // VERSION is current ek package version
-const VERSION = "13.0.1"
+const VERSION = "13.1.0"
 
 // ////////////////////////////////////////////////////////////////////////////////// //
